ca-ticket-agent/requirements.txt
playwright==1.49.1
openpyxl==3.1.5
pyyaml==6.0.2
requests==2.32.3
pandas==2.2.3
